pyproject.toml
[project]
name = "code-splitter-poetic"
version = "0.1.0"
description = ""
authors = [
    {name = "erbaz",email = "erbazkamran@gmail.com"}
]
requires-python = "<4.0,>=3.9"
dependencies = [
    "llama-index (>=0.13.1,<0.14.0)",
    "tree-sitter-language-pack (>=0.9.0,<0.10.0)",
    "llama-index-core (>=0.13.1,<0.14.0)",
    "llama-index-vector-stores-milvus (>=0.9.0,<0.10.0)",
    "pymilvus (>=2.6.0,<3.0.0)",
    "llama-index-embeddings-gemini (>=0.4.0,<0.5.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "llama-index-llms-gemini (>=0.6.0,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
